split Card.jsx in differents files
