+++ packages/vulcan-core/lib/modules/components.js
@@ -5,7 +5,6 @@ Also imported by Storybook
 */
 
 import './components/Avatar.jsx';
-import './components/Card.jsx';
 import './components/Loading.jsx';
 import './components/LoadingButton.jsx';
 import './components/NewButton.jsx';
