+++ packages/vulcan-core/lib/modules/components/Card/CardItemHTML.jsx
@@ -1,0 +1,8 @@
+import { registerComponent } from 'meteor/vulcan:lib';
+import React from 'react';
+
+// HTML
+const CardItemHTML = ({ value }) => (
+  <div className="contents-html" dangerouslySetInnerHTML={{ __html: value }} />
+);
+registerComponent({ name: 'CardItemHTML', component: CardItemHTML });
